Fix button drawing on startup
A number of issues here.
buid_buttons is called for a specific tab, however, it was redrawing it
for whatever is the active tab.
I.e. it kept on redrawing it for no reason.
Also, the initial tab was not set yet on startup, causing icons / text
not to show up on startup.

If the first highlighted button happened to have text, that would
trigger an update and why it "worked".

Removing the tab show/hide. It's not clear why that is there - perhaps
forcing a repaint for some reason. Will see if any issues surface.