Note that logout/login is necessary for input.
With the uaccess tag in the udev rules, interaction is needed between systemd-logind, PAM and dbus to set the device permissions.  

This happens at login time, so at initial setup, a login/logout will be needed to get keypresses working.  Display works right away so it looks like steamdeck-ui is working fine but the buttons don't work.  Logging out and in fixes this.  

cf. https://unix.stackexchange.com/questions/467382/udev-uaccess-and-hid